Move DrawText trait to kas::draw
